feat: token price + monetary values now support decimals
Token price like \$0.05 / α or \$5.25 / α is now accepted (and required
for realistic alpha pricing — most tokens trade in sub-dollar territory).

Backend:
- queries.set_settlement_distribution(token_price_usd: float) — drops
  int() coercion; math runs in float, results rounded to 2 decimals
  for cents precision. SQLite type affinity stores floats fine into
  the legacy INTEGER columns.
- queries.preview_kas_distribution(amount_usd: float) — same shift.
- kas_totals returns floats now (balance/contributed/distributed).
- create_kas_distribution accepts float amount.

API:
- Pydantic SettlementDistributionBody.token_price_usd: float
- Pydantic KasDistributionBody.amount_usd: float
- /api/kas/preview ?amount_usd= now accepts float

Display:
- _format_usd(value, decimals=2): handles floats, hides trailing .00 when
  the value is whole (\$5 not \$5.00) and shows up to N decimals otherwise.
- Token price KPI uses `| usd(4)` so prices like \$0.0035 display fully.
- All other monetary fields use default 2 decimals.

JS:
- _parseUsd no longer rounds — returns float so "5.25", "0.05", "1.5k"
  (= 1500) all preserve precision.
- _fmtUsd / _usd helpers updated to match Python formatter behavior.

Tests:
- Existing assertions updated from integer-exact to pytest.approx where
  rounding causes ±0.01 differences. 150 tests pass.

Manual e2e:
  token_price=\$0.05 × 45.73 α = \$2.29 → reward \$0.68 + kas \$1.61 ✓
  token_price=\$5.25 × 45.73 α = \$240.07 → reward \$72.03 + kas \$168.05 ✓

+++ src/emission_tracker/web/routes_api.py
@@ -17,11 +17,11 @@ class SettlementCreateBody(BaseModel):
 
 
 class SettlementDistributionBody(BaseModel):
-    token_price_usd: int
+    token_price_usd: float
 
 
 class KasDistributionBody(BaseModel):
-    amount_usd: int
+    amount_usd: float
     note: str | None = None
 
 
@@ -179,7 +179,7 @@ def get_kas_balance(request: Request):
 
 
 @router.get("/kas/preview")
-def preview_kas(request: Request, amount_usd: int = Query(ge=0)):
+def preview_kas(request: Request, amount_usd: float = Query(ge=0)):
     """Read-only preview of how `amount_usd` would split across all-time
     contributors. Useful for the Distribusi kas form before confirming."""
     try:
